Update Drop_Neckfinder.ipynb
Massive overhaul.  Removed contour finding completely and replaced it with a column-by-column search of arrays.  This was made possible due to finding out how to reduce the loaded video file to a single channel, which was much easier to manipulate.

Leading front detection is vastly improved.  Now instead of excluding part of the image, the function checks the minima it finds to verify that the drop is present on both sides of it.  If not, it indicates that a leading edge has been found; the function then drops that datapoint and repeats until it finds an acceptable neck.